# river-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the River self-hosted media platform.\nAuth flow: POST /api/auth/login → use the returned\naccess_token as a Bearer token on the protected routes.\nThe stream_token is a longer-lived JWT restricted to\n/stream and /download endpoints (the \u003cvideo\u003e element\ncan't send custom auth headers; it goes in the URL as\n?token=…).",
        "title": "River API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/admin/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Upload media file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Media file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "movie | episode",
                        "name": "type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target library",
                        "name": "library_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Used as folder/file name",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Season number (episodes only)",
                        "name": "season",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Episode number (episodes only)",
                        "name": "episode",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "{library_id, path}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/watchlist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchlist"
                ],
                "summary": "List watchlist",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchlist"
                ],
                "summary": "Add to watchlist",
                "parameters": [
                    {
                        "description": "Media reference",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.watchlistAddRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.WatchlistItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/watchlist/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "watchlist"
                ],
                "summary": "Remove from watchlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Watchlist entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.watchlistAddRequest": {
            "type": "object",
            "required": [
                "media_id",
                "media_type"
            ],
            "properties": {
                "media_id": {
                    "type": "string"
                },
                "media_type": {
                    "type": "string",
                    "enum": [
                        "movie",
                        "tvshow",
                        "audiobook"
                    ]
                }
            }
        },
...
        "models.WatchlistItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mediaID": {
                    "type": "string"
                },
                "mediaType": {
                    "description": "\"movie\" | \"tvshow\" | \"audiobook\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT — paste \"Bearer \u003ctoken\u003e\" or use the Authorize button.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}